create seperate BeakCLI target
