+++ Sources/BeakCLI/BeakOptions.swift
@@ -1,0 +1,12 @@
+import PathKit
+
+public struct BeakOptions {
+
+    public let cachePath: Path
+    public let packageName: String
+
+    public init(cachePath: Path = "~/.beak/builds", packageName: String = "BeakFile") {
+        self.cachePath = cachePath.normalize()
+        self.packageName = packageName
+    }
+}
