+++ Sources/BeakCLI/Commands/EditCommand.swift
@@ -1,5 +1,6 @@
 import PathKit
 import SwiftCLI
+import BeakCore
 
 class EditCommand: BeakCommand {
