+++ Sources/BeakCLI/Commands/FunctionCommand.swift
@@ -1,5 +1,6 @@
 import PathKit
 import SwiftCLI
+import BeakCore
 
 class FunctionCommand: BeakCommand {
