+++ Sources/BeakCLI/Commands/ListCommand.swift
@@ -1,5 +1,6 @@
 import PathKit
 import SwiftCLI
+import BeakCore
 
 class ListCommand: BeakCommand {
